fix: prevent clickConfirm/clickCancel from failing if there's no popup
