+++ src/staticMethods/dom.js
@@ -28,9 +28,9 @@ export const isVisible = () => {
 /*
  * Global function to click 'Confirm' button
  */
-export const clickConfirm = () => dom.getConfirmButton().click()
+export const clickConfirm = () => dom.getConfirmButton() && dom.getConfirmButton().click()
 
 /*
  * Global function to click 'Cancel' button
  */
-export const clickCancel = () => dom.getCancelButton().click()
+export const clickCancel = () => dom.getCancelButton() && dom.getCancelButton().click()
