+++ test/qunit/tests.js
@@ -674,22 +674,6 @@ QUnit.test('timer', (assert) => {
   })
 })
 
-QUnit.test('confirm button', (assert) => {
-  const done = assert.async()
-  Swal.fire({
-    input: 'radio',
-    inputOptions: {
-      'one': 'one',
-      'two': 'two'
-    }
-  }).then((result) => {
-    assert.deepEqual(result, { value: 'two' })
-    done()
-  })
-  $('.swal2-radio').querySelector('input[value="two"]').checked = true
-  Swal.clickConfirm()
-})
-
 QUnit.test('params validation', (assert) => {
   assert.ok(Swal.isValidParameter('title'))
   assert.notOk(Swal.isValidParameter('foobar'))
